+++ tests/testthat/test-novax.R
@@ -44,7 +44,8 @@ test_that("novax_equilib returns equlibrium conditions", {
 test_that("novax_baseline returns baseline conditions", {
 
   y <- novax_baseline(1:3, 1:11)
-  expect_equal(names(y), c("incid", "cum_incid", "cum_diag_a", "cum_diag_s"))
+  expect_equal(names(y), c("incid", "cum_incid", "cum_diag_a",
+                           "cum_diag_s", "cum_screened"))
   expect_equal(dim(y$incid), c(11, 3))
   expect_equal(dim(y$cum_incid), c(11, 3))
   expect_equal(colSums(y$incid), y$cum_incid[11, ])
@@ -52,5 +53,5 @@ test_that("novax_baseline returns baseline conditions", {
   y2 <- novax_baseline(1:100, t = 10)
   expect_equal(length(y2[[1]]), 100)
 
-  expect_error(novax_baseline(1, 21), "t must be an integer between 1 and 20")
+  expect_error(novax_baseline(1, 31), "t must be an integer between 1 and 30")
 })